# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS portfolio part2",
        "version": "0.1"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "summary": "Finds the list of all books ",
                "operationId": "1ea73ebde846542f75f27f102baf067f",
                "responses": {
                    "200": {
                        "description": "Browse the list of all books"
                    }
                }
            },
            "post": {
                "summary": "Store new book",
                "operationId": "f793f459b06e946f54c04166bc76187f",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the book",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "subtitle",
                        "in": "query",
                        "description": "Subtitle of the book",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Store new books"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "summary": "Finds Books by book id",
                "operationId": "d946c9001412fe4bbfa69dc638734745",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id to filter by",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show selected Book"
                    },
                    "404": {
                        "description": "Show error when Book not Found"
                    }
                }
            },
            "delete": {
                "summary": "Delete the book",
                "operationId": "6e946588a5159f78ff5b31dec61300e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id to filter by",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete the book"
                    },
                    "404": {
                        "description": "Show error when book not Found"
                    }
                }
            },
            "patch": {
                "summary": "Update the book",
                "operationId": "ec90c97426d8e9b709dc5c78cb29ba1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the book",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the book",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "subtitle",
                        "in": "query",
                        "description": "Subtitle of the book",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update the book"
                    },
                    "404": {
                        "description": "Show error when book not Found"
                    }
                }
            }
        },
        "/api/genres": {
            "get": {
                "summary": "Finds the list of all genres ",
                "operationId": "8af3149ee50579f5473f32debd2efc4f",
                "responses": {
                    "200": {
                        "description": "Browse the list of all genres"
                    }
                }
            },
            "post": {
                "summary": "Store new genre",
                "operationId": "9c2dd28e55b823a3699b14f174d8d894",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name of the genre",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description of the genre",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Store new genre"
                    }
                }
            }
        },
        "/api/genres/{id}": {
            "get": {
                "summary": "Finds genre by genre id",
                "operationId": "0dc073d27e7e2049c9826d8fc093dc06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Genre id to filter by",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show selected genre"
                    },
                    "404": {
                        "description": "Show error when genre not Found"
                    }
                }
            },
            "delete": {
                "summary": "Delete the genre",
                "operationId": "c2022853f5c37c0dc6384e863918b7de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Genre id to filter by",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete the genre"
                    },
                    "404": {
                        "description": "Show error when genre not Found"
                    }
                }
            },
            "patch": {
                "summary": "Update the genre",
                "operationId": "e0ce3e0d5b509ae71079691fe262f564",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the genre",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name of the genre",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description of the genre",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update the genre"
                    },
                    "404": {
                        "description": "Show error when genre not Found"
                    }
                }
            }
        },
...
    }
}